⏪ Schema: Switch back to ID as int
Yeah, lots of bad things happened with type conversion errors, etc. Instead, I made the right choice and set the `toml_metadata` method on the JamfScript class to exclude attributes with `none` as value. Tomlkit doesn't need to serialise it if it is not present in the dict.

It also means new users won't be tempted to assign it an ID.

Of course this does mean we need to explicitly handle the missing attribute when using `local.get_script_by_id`: On iteration, we ignore items with None as ID. Easy.