⚰️ Remove `push_from_metadata`
The script evolved around this and now it looks dumb.

CLI friendly error handling should be implemented elsewhere. It's now handled by the `scripts` module.

And it looks real silly to have a function pass in a class instance and an arg for the class's method...

Unnecessary abstraction and fake dependency injection. Yuck.