🐛 Treat script extensions as part of filename
The reason for this is the shebang determines execution. The script extension is completely unnecessary. Tacking ".sh" on the end of every script was highly presumptuous (of course I didn't realise how many Applescripts, Zsh and Python scripts we had).

If someone elects to put ".sh" in the new script editor title in Jamf or in a local script to push with Jast, so be it. We will treat it as part of the filename. Metadata files too, will ignore it, so you may end up with metadata filenames like "MyScript.sh.toml".